+++ backend/app.js
@@ -1,67 +1,12 @@
 const express = require("express");
 const multer = require("multer");
 const db = require("./database");
-const parseIcal = require("./services/icalParserService");
 
 const app = express();
-const upload = multer({ storage: multer.memoryStorage() });
-
-app.post("/upload", upload.single("file"), (req, res) => {
-  if (!req.file) {
-    return res.status(400).send("No file uploaded.");
-  }
-
-  try {
-    // Parse the .ics file
-    const icsData = req.file.buffer.toString("utf-8");
-    console.log("ICS Data:", icsData); // Log raw ICS data
-    const events = parseIcal(icsData);
-    console.log("Parsed Events:", events); // Log parsed events
-
-    // Save events to the database
-    const insertPromises = events.map((event) => {
-      return new Promise((resolve, reject) => {
-        const query = `
-          INSERT INTO Schedules (user_id, name, description, due_date, category, priority, labels, start_time, end_time, status, source)
-          VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-        `;
-        const params = [
-          1, // Replace with actual user_id if needed
-          event.name,
-          event.description,
-          null, // Replace with actual due_date if needed
-          'work',
-          event.priority,
-          event.labels,
-          event.start_time,
-          event.end_time,
-          event.status,
-          "ics",
-        ];
-
-        db.run(query, params, function (err) {
-          if (err) {
-            console.error("Database Insert Error:", err.message);
-            reject(err);
-          } else {
-            resolve();
-          }
-        });
-      });
-    });
-
-    Promise.all(insertPromises)
-      .then(() => res.status(200).send("Events successfully saved."))
-      .catch((error) => res.status(500).send("Error processing file."));
-  } catch (error) {
-    console.error("File Processing Error:", error);
-    res.status(500).send("Error processing file.");
-  }
-});
-
 app.use(express.json());
 
 // Load route files
+const uploadRoutes = require('./routes/upload'); // upload ics file
 const usersRoutes = require("./routes/users"); // Ensure correct path
 const preferencesRoutes = require("./routes/preferences"); // Ensure correct path
 const plansRoutes = require("./routes/plans"); // Ensure correct path
@@ -72,6 +17,7 @@ const generatescheduleRoutes = require("./routes/generateschedule"); // Ensure c
 const generatesummaryRoutes = require("./routes/generatesummary"); // Ensure correct path
 
 // Use the routes as middleware
+app.use('/upload', uploadRoutes);
 app.use("/users", usersRoutes);
 app.use("/preferences", preferencesRoutes);
 app.use("/plans", plansRoutes);
