+++ backend/sampleics.ics
@@ -1,188 +1,98 @@
 BEGIN:VCALENDAR
 VERSION:2.0
-PRODID:-//Sample ICS for Work Week//EN
+PRODID:-//Sample Calendar for Sept-Oct 2024//EN
+
+BEGIN:VTIMEZONE
+TZID:UTC
+END:VTIMEZONE
 
 BEGIN:VEVENT
-UID:meeting-1@sample.com
-DTSTAMP:20240918T090000Z
-DTSTART:20240918T090000Z
-DTEND:20240918T100000Z
-SUMMARY:Team Meeting
-DESCRIPTION:Weekly status meeting with the team.
-LOCATION:Meeting Room 1
-STATUS:CONFIRMED
-END:VEVENT
-
-BEGIN:VEVENT
-UID:focus-time-1@sample.com
-DTSTAMP:20240918T103000Z
-DTSTART:20240918T103000Z
-DTEND:20240918T120000Z
-SUMMARY:Focus Time - Project Work
-DESCRIPTION:Dedicated time for working on the project.
-STATUS:CONFIRMED
-END:VEVENT
-
-BEGIN:VEVENT
-UID:break-1@sample.com
-DTSTAMP:20240918T120000Z
-DTSTART:20240918T120000Z
-DTEND:20240918T123000Z
-SUMMARY:Lunch Break
-DESCRIPTION:Lunch time to recharge.
-STATUS:CONFIRMED
-END:VEVENT
-
-BEGIN:VEVENT
-UID:client-call-1@sample.com
-DTSTAMP:20240918T130000Z
-DTSTART:20240918T130000Z
-DTEND:20240918T140000Z
-SUMMARY:Client Call
-DESCRIPTION:Monthly call with the client.
-LOCATION:Zoom
-STATUS:CONFIRMED
-END:VEVENT
-
-BEGIN:VEVENT
-UID:focus-time-2@sample.com
-DTSTAMP:20240918T143000Z
-DTSTART:20240918T143000Z
-DTEND:20240918T163000Z
-SUMMARY:Focus Time - Code Review
-DESCRIPTION:Reviewing code and optimizing solutions.
-STATUS:CONFIRMED
-END:VEVENT
-
-BEGIN:VEVENT
-UID:daily-wrapup-1@sample.com
-DTSTAMP:20240918T163000Z
-DTSTART:20240918T163000Z
-DTEND:20240918T170000Z
-SUMMARY:Daily Wrap-Up
-DESCRIPTION:End-of-day wrap-up and notes for tomorrow.
-STATUS:CONFIRMED
-END:VEVENT
-
-BEGIN:VEVENT
-UID:meeting-2@sample.com
-DTSTAMP:20240919T090000Z
-DTSTART:20240919T090000Z
-DTEND:20240919T100000Z
-SUMMARY:Planning Session
-DESCRIPTION:Planning tasks for the upcoming sprint.
-LOCATION:Zoom
-STATUS:CONFIRMED
-END:VEVENT
-
-BEGIN:VEVENT
-UID:focus-time-3@sample.com
-DTSTAMP:20240919T103000Z
-DTSTART:20240919T103000Z
-DTEND:20240919T120000Z
-SUMMARY:Focus Time - Research
-DESCRIPTION:Time for research and analysis.
-STATUS:CONFIRMED
-END:VEVENT
-
-BEGIN:VEVENT
-UID:break-2@sample.com
-DTSTAMP:20240919T120000Z
-DTSTART:20240919T120000Z
-DTEND:20240919T123000Z
-SUMMARY:Lunch Break
-DESCRIPTION:Lunch break to recharge.
-STATUS:CONFIRMED
-END:VEVENT
-
-BEGIN:VEVENT
-UID:demo-preparation@sample.com
-DTSTAMP:20240919T130000Z
-DTSTART:20240919T130000Z
-DTEND:20240919T140000Z
-SUMMARY:Demo Preparation
-DESCRIPTION:Preparing for the upcoming product demo.
-STATUS:CONFIRMED
-END:VEVENT
-
-BEGIN:VEVENT
-UID:focus-time-4@sample.com
-DTSTAMP:20240919T143000Z
-DTSTART:20240919T143000Z
-DTEND:20240919T163000Z
-SUMMARY:Focus Time - Bug Fixing
-DESCRIPTION:Fixing critical bugs from recent feedback.
-STATUS:CONFIRMED
-END:VEVENT
-
-BEGIN:VEVENT
-UID:daily-wrapup-2@sample.com
-DTSTAMP:20240919T163000Z
-DTSTART:20240919T163000Z
-DTEND:20240919T170000Z
-SUMMARY:Daily Wrap-Up
-DESCRIPTION:End-of-day wrap-up and notes for tomorrow.
-STATUS:CONFIRMED
-END:VEVENT
-
-BEGIN:VEVENT
-UID:meeting-3@sample.com
+UID:weekly-team-meeting@sample.com
 DTSTAMP:20240920T090000Z
-DTSTART:20240920T090000Z
-DTEND:20240920T100000Z
-SUMMARY:Stakeholder Update
-DESCRIPTION:Updating stakeholders on project status.
+DTSTART;TZID=UTC:20240923T090000
+DTEND;TZID=UTC:20240923T100000
+RRULE:FREQ=WEEKLY;UNTIL=20241020T235900Z;BYDAY=MO
+SUMMARY:Weekly Team Meeting
+DESCRIPTION:Regular team sync-up.
+LOCATION:Conference Room A
+STATUS:CONFIRMED
+END:VEVENT
+
+BEGIN:VEVENT
+UID:focus-time-morning@sample.com
+DTSTAMP:20240920T100000Z
+DTSTART;TZID=UTC:20240920T100000
+DTEND;TZID=UTC:20240920T120000
+RRULE:FREQ=DAILY;UNTIL=20241020T235900Z;BYDAY=MO,TU,WE,TH,FR
+SUMMARY:Morning Focus Time
+DESCRIPTION:Dedicated time for project work.
+STATUS:CONFIRMED
+END:VEVENT
+
+BEGIN:VEVENT
+UID:break-lunch@sample.com
+DTSTAMP:20240920T120000Z
+DTSTART;TZID=UTC:20240920T120000
+DTEND;TZID=UTC:20240920T123000
+RRULE:FREQ=DAILY;UNTIL=20241020T235900Z;BYDAY=MO,TU,WE,TH,FR
+SUMMARY:Lunch Break
+DESCRIPTION:Time to recharge and eat lunch.
+STATUS:CONFIRMED
+END:VEVENT
+
+BEGIN:VEVENT
+UID:client-call@sample.com
+DTSTAMP:20240924T130000Z
+DTSTART;TZID=UTC:20240924T130000
+DTEND;TZID=UTC:20240924T140000
+RRULE:FREQ=WEEKLY;UNTIL=20241020T235900Z;BYDAY=TU
+SUMMARY:Client Call
+DESCRIPTION:Weekly call with the client.
 LOCATION:Zoom
 STATUS:CONFIRMED
 END:VEVENT
 
 BEGIN:VEVENT
-UID:focus-time-5@sample.com
-DTSTAMP:20240920T103000Z
-DTSTART:20240920T103000Z
-DTEND:20240920T120000Z
-SUMMARY:Focus Time - Documentation
-DESCRIPTION:Writing technical documentation.
+UID:focus-time-afternoon@sample.com
+DTSTAMP:20240920T140000Z
+DTSTART;TZID=UTC:20240920T140000
+DTEND;TZID=UTC:20240920T170000
+RRULE:FREQ=DAILY;UNTIL=20241020T235900Z;BYDAY=MO,TU,WE,TH,FR
+SUMMARY:Afternoon Focus Time
+DESCRIPTION:Work on tasks and projects.
 STATUS:CONFIRMED
 END:VEVENT
 
 BEGIN:VEVENT
-UID:break-3@sample.com
-DTSTAMP:20240920T120000Z
-DTSTART:20240920T120000Z
-DTEND:20240920T123000Z
-SUMMARY:Lunch Break
-DESCRIPTION:Lunch break to recharge.
+UID:project-demo@sample.com
+DTSTAMP:20240930T150000Z
+DTSTART;TZID=UTC:20240930T150000
+DTEND;TZID=UTC:20240930T160000
+SUMMARY:Mid-Month Project Demo
+DESCRIPTION:Presenting project progress to stakeholders.
+LOCATION:Conference Room B
 STATUS:CONFIRMED
 END:VEVENT
 
 BEGIN:VEVENT
-UID:focus-time-6@sample.com
-DTSTAMP:20240920T130000Z
-DTSTART:20240920T130000Z
-DTEND:20240920T140000Z
-SUMMARY:Focus Time - Testing
-DESCRIPTION:Performing unit and integration testing.
+UID:evening-meeting@sample.com
+DTSTAMP:20240926T210000Z
+DTSTART;TZID=UTC:20240926T210000
+DTEND;TZID=UTC:20240926T213000
+SUMMARY:Evening Strategy Meeting
+DESCRIPTION:Discussing strategies for upcoming challenges.
+LOCATION:Zoom
 STATUS:CONFIRMED
 END:VEVENT
 
 BEGIN:VEVENT
-UID:focus-time-7@sample.com
-DTSTAMP:20240920T143000Z
-DTSTART:20240920T143000Z
-DTEND:20240920T163000Z
-SUMMARY:Focus Time - Feature Development
-DESCRIPTION:Developing new features for the product.
-STATUS:CONFIRMED
-END:VEVENT
-
-BEGIN:VEVENT
-UID:daily-wrapup-3@sample.com
-DTSTAMP:20240920T163000Z
-DTSTART:20240920T163000Z
-DTEND:20240920T170000Z
-SUMMARY:Daily Wrap-Up
-DESCRIPTION:End-of-day wrap-up and notes for tomorrow.
+UID:end-of-month-review@sample.com
+DTSTAMP:20241020T150000Z
+DTSTART;TZID=UTC:20241020T150000
+DTEND;TZID=UTC:20241020T160000
+SUMMARY:End-of-Month Review
+DESCRIPTION:Reviewing achievements and setting goals for the next period.
+LOCATION:Conference Room A
 STATUS:CONFIRMED
 END:VEVENT
